Add camera timeout parameter

+++ gst-zedxone-src/gstzedxonesrc.h
@@ -50,7 +50,7 @@ struct _GstZedXOneSrc {
     gint camera_resolution;   // Camera resolution [enum]
     gint camera_fps;          // Camera FPS
     gint verbose_level;
-    guint16 cam_timeout_msec;
+    gint cam_timeout_msec;
     gint camera_id;
     //gint camera_sn;
     gboolean swap_rb; //swap for RGB(A) or BGR(A) output
